Fix for image rendering in for loops
When rendering more than 2 images in a report, only the 2 first would appear.
This is because Libreoffice does not let 2 images have the same name.
Before this fix, the first image was generated with name='Aeroo Image '
and the second with name='Image2'. Any other image was not rendered.

If we let the attribute empty, Libreoffice generates a name automatically for
each image which solves the problem.